Lógica backend sincronizada con la base de datos funcional

+++ src/main/java/com/biblioteca/sistemabiblioteca/controller/UsuarioController.java
@@ -9,12 +9,12 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 /**
- * CONTROLADOR DE GESTIÓN DE USUARIOS
- * Centraliza las operaciones de administración de cuentas para alumnos y docentes (RF-01 al RF-04).
+ * CONTROLADOR DE GESTIÓN DE USUARIOS - ACTUALIZADO PARA biblioteca_universidad
+ * Centraliza las operaciones de administración de cuentas (RF-01 al RF-04).
  */
 @RestController
 @RequestMapping("/api/usuarios")
-@CrossOrigin(origins = "*") // Clave para permitir peticiones desde el frontend en JavaFX.
+@CrossOrigin(origins = "*") 
 public class UsuarioController {
 
     @Autowired
@@ -25,18 +25,14 @@ public class UsuarioController {
 
     /**
      * RF-01: REGISTRO DE USUARIOS
-     * Procesa la creación de nuevos usuarios delegando la validación de reglas 
-     * (como correo único o formato de datos) al componente de Servicio.
      */
     @PostMapping
     public Usuarios crearUsuario(@RequestBody Usuarios usuario) {
-        // El Service se encarga de que no se dupliquen correos y de inicializar el estado.
         return usuarioService.registrarUsuario(usuario);
     }
 
     /**
      * RF-02: CONSULTA GENERAL DE USUARIOS
-     * Retorna el listado completo de usuarios registrados para su visualización en tablas o reportes.
      */
     @GetMapping
     public List<Usuarios> listarUsuarios() {
@@ -45,16 +41,15 @@ public class UsuarioController {
 
     /**
      * RF-04: ELIMINACIÓN LÓGICA (DESACTIVACIÓN)
-     * En lugar de eliminar la fila de la base de datos, cambiamos el atributo 'estado'.
-     * Esto mantiene la integridad referencial si el usuario tiene préstamos históricos.
+     * Actualizado: El ID ahora se maneja como Integer para ser compatible con el script SQL.
      */
     @PutMapping("/{id}/desactivar")
-    public Usuarios desactivarUsuario(@PathVariable Long id) {
-        // Verificamos la existencia del usuario antes de proceder con el cambio de estado.
+    public Usuarios desactivarUsuario(@PathVariable Integer id) { // CAMBIO: Long -> Integer
+        // Verificamos la existencia usando la nueva llave primaria id_usuario (mapeada en el modelo)
         Usuarios usuario = usuarioRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Usuario con ID " + id + " no existe"));
         
-        // Marcamos el estado como inactivo (false) para restringir sus privilegios sin borrar sus datos.
+        // Cambio de estado para restricción de privilegios sin pérdida de datos históricos.
         usuario.setEstado(false); 
         return usuarioRepository.save(usuario);
     }
